Update/Disable regression test due to compact source file feature
JavaRefactoringActionsProviderTest#test190101:

 - javac 25 assumes everything without a class declaration is a compact
   source file -> disabled

JavaCompletionTaskXXXFeaturesTest:

 - Records (final) are a JDK 16+ feature. Reflecting this in the tests.
 - moving into the JDk 16 completion tests, using the appropriate
   source level.
