--- java/java.completion/test/unit/data/goldenfiles/org/netbeans/modules/java/completion/JavaCompletionTaskTest/1.8/typesRecordStaticMembersAndVars.pass
@@ -1,6 +1,0 @@
-Readable
-ReflectiveOperationException
-Runnable
-Runtime
-RuntimeException
-RuntimePermission
