+++ java/java.completion/test/unit/data/goldenfiles/org/netbeans/modules/java/completion/JavaCompletionTaskTest/14/typesRecordStaticMembersAndVars.pass
@@ -1,5 +1,6 @@
 Readable
 Record
+Records
 ReflectiveOperationException
 Runnable
 Runtime
